refactor: renamed file signable-token to tokens
